# views/docs/swagger.yaml
basePath: /v3
definitions:
  action_log.Request:
    properties:
      entry:
        example: Changed Preferred OIs to RP
        type: string
    required:
    - entry
    type: object
  action_log.Response:
    properties:
      cid:
        example: 1293257
        type: integer
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      created_by:
        example: '''1234567'' or ''System'''
        type: string
      entry:
        example: Changed Preferred OIs to RP
        type: string
      id:
        example: 1
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      updated_by:
        example: '''1234567'' or ''System'''
        type: string
    type: object
...
  constants.FacilityID:
    enum:
    - ZAE
    - ZHQ
    - ZZN
    - ZZI
    - ZAB
    - ZAN
    - ZTL
    - ZBW
    - ZAU
    - ZOB
    - ZDV
    - ZFW
    - HCF
    - ZHU
    - ZID
    - ZJX
    - ZKC
    - ZLA
    - ZME
    - ZMA
    - ZMP
    - ZNY
    - ZOA
    - ZLC
    - ZSE
    - ZDC
    type: string
    x-enum-varnames:
    - AcademyFacility
    - HeadquartersFacility
    - NonMemberFacility
    - InactiveFacility
    - AlbuquerqueFacility
    - AnchorageFacility
    - AtlantaFacility
    - BostonFacility
    - ChicagoFacility
    - ClevelandFacility
    - DenverFacility
    - FortWorthFacility
    - HonoluluFacility
    - HoustonFacility
    - IndianapolisFacility
    - JacksonvilleFacility
    - KansasCityFacility
    - LosAngelesFacility
    - MemphisFacility
    - MiamiFacility
    - MinneapolisFacility
    - NewYorkFacility
    - OaklandFacility
    - SaltLakeFacility
    - SeattleFacility
    - WashingtonFacility
...
  disciplinary_log.Request:
    properties:
      entry:
        example: Misconduct in discord
        type: string
      vatusa_only:
        example: true
        type: boolean
    required:
    - entry
    type: object
  disciplinary_log.Response:
    properties:
      cid:
        example: 1293257
        type: integer
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      created_by:
        example: '''1234567'' or ''System'''
        type: string
      entry:
        example: Changed Preferred OIs to RP
        type: string
      id:
        example: 1
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      updated_by:
        example: '''1234567'' or ''System'''
        type: string
      vatusa_only:
        example: true
        type: boolean
    type: object
...
  facility_log.Request:
    properties:
      entry:
        example: Changed Preferred OIs to RP
        type: string
      facility:
        allOf:
        - $ref: '#/definitions/constants.FacilityID'
        example: ZDV
    required:
    - entry
    - facility
    type: object
  facility_log.Response:
    properties:
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      created_by:
        example: '''1234567'' or ''System'''
        type: string
      entry:
        example: Change URL from 'denartcc.org' to 'zdvartcc.org'
        type: string
      facility:
        allOf:
        - $ref: '#/definitions/constants.FacilityID'
        example: ZDV
      id:
        example: 1
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      updated_by:
        example: '''1234567'' or ''System'''
        type: string
    type: object
...
  utils.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
info:
  contact:
    email: vatusa6@vatusa.net
    name: VATUSA Support
    url: http://www.swagger.io/support
  description: VATUSAs public API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VATUSA API
  version: "0.1"
paths:
...
  /facility/{FacilityID}/log:
    get:
      consumes:
      - application/json
      description: List facility log entries
      parameters:
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/facility_log.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: List facility log entries
      tags:
      - facility-log
    post:
      consumes:
      - application/json
      description: Create a new facility log entry
      parameters:
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      - description: Facility Log Entry
        in: body
        name: facility_log
        required: true
        schema:
          $ref: '#/definitions/facility_log.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/facility_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create a new facility log entry
      tags:
      - facility-log
  /facility/{FacilityID}/log/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a facility log entry
      parameters:
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      - description: Facility Log Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Delete a facility log entry
      tags:
      - facility-log
    patch:
      consumes:
      - application/json
      description: Patch a facility log entry
      parameters:
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      - description: Facility Log Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Facility Log Entry
        in: body
        name: facility_log
        required: true
        schema:
          $ref: '#/definitions/facility_log.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/facility_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Patch a facility log entry
      tags:
      - facility-log
    put:
      consumes:
      - application/json
      description: Update a facility log entry
      parameters:
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      - description: Facility Log Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Facility Log Entry
        in: body
        name: facility_log
        required: true
        schema:
          $ref: '#/definitions/facility_log.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/facility_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update a facility log entry
      tags:
      - facility-log
...
  /user/{cid}/action-log:
    get:
      consumes:
      - application/json
      description: List all action log entries
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/action_log.Response'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get all action log entries
      tags:
      - action-log
    post:
      consumes:
      - application/json
      description: Create a new action log entry
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Action Log Entry
        in: body
        name: action_log
        required: true
        schema:
          $ref: '#/definitions/action_log.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/action_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create a new action log entry
      tags:
      - action-log
  /user/{cid}/action-log/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an action log entry
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Action Log Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Delete an action log entry
      tags:
      - action-log
    patch:
      consumes:
      - application/json
      description: Patch an action log entry
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Action Log Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Action Log Entry
        in: body
        name: action_log
        required: true
        schema:
          $ref: '#/definitions/action_log.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Patch an action log entry
      tags:
      - action-log
    put:
      consumes:
      - application/json
      description: Update an action log entry
      parameters:
      - description: Action Log Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Action Log Entry
        in: body
        name: action_log
        required: true
        schema:
          $ref: '#/definitions/action_log.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update an action log entry
      tags:
      - action-log
  /user/{cid}/disciplinary-log:
    get:
      consumes:
      - application/json
      description: List all disciplinary log entries
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: VATUSA Only
        in: query
        name: vatusa_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/disciplinary_log.Response'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get all disciplinary log entries
      tags:
      - disciplinary-log
    post:
      consumes:
      - application/json
      description: Create a new disciplinary log entry
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Disciplinary Log Entry
        in: body
        name: action_log
        required: true
        schema:
          $ref: '#/definitions/disciplinary_log.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/disciplinary_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create a new disciplinary log entry
      tags:
      - disciplinary-log
  /user/{cid}/disciplinary-log/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a disciplinary log entry
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Disciplinary Log Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Delete a disciplinary log entry
      tags:
      - disciplinary-log
    patch:
      consumes:
      - application/json
      description: Patch an disciplinary log entry
      parameters:
      - description: User CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Disciplinary Log Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Disciplinary Log Entry
        in: body
        name: disciplinary_log
        required: true
        schema:
          $ref: '#/definitions/disciplinary_log.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/disciplinary_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Patch an disciplinary log entry
      tags:
      - disciplinary-log
    put:
      consumes:
      - application/json
      description: Update a disciplinary log entry
      parameters:
      - description: Disciplinary Log Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Disciplinary Log Entry
        in: body
        name: disciplinary_log
        required: true
        schema:
          $ref: '#/definitions/disciplinary_log.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/disciplinary_log.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update a disciplinary log entry
      tags:
      - disciplinary-log
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"